Experimenting with thresholding and PiCamera
New script which messes around with HSV thresholding and properties of the Pi Camera, specifically adjusting exposure and auto white balance gains. The former I know works well, the latter needs to be tested on actual hardware. I need to see if I can adjust these values on the go while we are still pulling frames and running the pipeline. Pushing this commit so I can clone it on my Pi.